Organize types
