+++ samples/MDiator.SampleApp/Events/OrderCreatedEvent.cs
@@ -1,0 +1,7 @@
+﻿namespace MDiator.SampleApp.Events
+{
+    public class OrderCreatedEvent : IMDiatorEvent
+    {
+        public int OrderId { get; set; }
+    }
+}
