+++ samples/MDiator.SampleApp/Program.cs
@@ -1,42 +1,16 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using MDiator.SampleApp.Events;
+using MDiator.SampleApp.Requests;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace MDiator.SampleApp
 {
-    public class CreateUserCommand : IMDiatorRequest<string>
-    {
-        public string UserName { get; set; }
-    }
-
-    public class CreateUserHandler : IMDiatorHandler<CreateUserCommand, string>
-    {
-        public Task<string> Handle(CreateUserCommand request)
-        {
-            return Task.FromResult($"✅ User '{request.UserName}' created.");
-        }
-    }
-
-    // ======== EVENT ==========
-    public class OrderCreatedEvent : IMDiatorEvent
-    {
-        public int OrderId { get; set; }
-    }
-
-    public class NotifyTeamHandler : IMDiatorEventHandler<OrderCreatedEvent>
-    {
-        public Task Handle(OrderCreatedEvent e)
-        {
-            Console.WriteLine($"🔔 Team notified about Order #{e.OrderId}");
-            return Task.CompletedTask;
-        }
-    }
-
-    internal class Program
+    public class Program
     {
         static async Task Main(string[] args)
         {
             // ✅ Setup DI
             var services = new ServiceCollection();
-            services.AddMDiator(typeof(Program).Assembly); // auto-register handlers
+            services.AddMDiator(typeof(Program).Assembly);
             var provider = services.BuildServiceProvider();
 
             var mediator = provider.GetRequiredService<IMediator>();
