+++ samples/MDiator.SampleApp/Requests/CreateUserCommand.cs
@@ -1,0 +1,7 @@
+﻿namespace MDiator.SampleApp.Requests
+{
+    public class CreateUserCommand : IMDiatorRequest<string>
+    {
+        public string UserName { get; set; }
+    }
+}
