+++ src/MDiator/Extensions/ServiceCollectionExtensions.cs
@@ -1,7 +1,7 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using MDiator;
 using System.Reflection;
 
-namespace MDiator
+namespace Microsoft.Extensions.DependencyInjection
 {
     public static class ServiceCollectionExtensions
     {
